[fix] Due to recent changes in 1337x bot detection, the current method of web scraping no longer works.
To counteract this I am replacing the networking requests module with
Selenium.

Current issues are that Selenium --headless=new does not generate to
correct HTML DOM to scrape. This is fine when not in headless mode but
why would anyone run that.

Abandoning the Python version of this and exploring different paths.